added mysql_metrics, mysql_user and mysql_passwd as variables for the default config file metrics, user and passwd respectively. Apologies if I'm overstepping, but I did not see where the default username and password was defined, so I defined them.
Also added parts to the mysql template that used the values for mysql_metrics, mysql_user and mysql_passwd unless there were overrides defined in the "servers" array.
Added the default that the "name" was the same as the "host" unless otherwise specified. I've seen most people set it up that way, so there's no need to clog up the node manifest with duplicate name and host entries. With the above changes, it's sufficient for folks who want to measure the same thing on all servers, who use the same username and password, and have one instance per host, to only define the "host" for each server...this helps a lot when you're
 monitoring 100 servers! (1 line instead of 5+ lines each).

Fixed typo in the newrelic.properties.erb file so the license key
actually shows up.